webapp/requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==2.0.1
Flask-Cors==3.0.10
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dotenv==0.18.0
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
tmdbv3api==1.7.6
toml==0.10.2
urllib3==1.25.8
uWSGI==2.0.21
webencodings==0.5.1
Werkzeug==2.0.1
